+++ src/Rule/Date/DateTime.php
@@ -16,24 +16,28 @@ use Yiisoft\Validator\WhenInterface;
  * @psalm-import-type WhenType from WhenInterface
  */
 #[Attribute(Attribute::TARGET_PROPERTY | Attribute::IS_REPEATABLE)]
-final class Date extends BaseDate
+final class DateTime extends BaseDate
 {
     /**
      * @psalm-param non-empty-string|null $timeZone
      * @psalm-param IntlDateFormatterFormat|null $dateType
+     * @psalm-param IntlDateFormatterFormat|null $timeType
      * @psalm-param IntlDateFormatterFormat|null $messageDateType
+     * @psalm-param IntlDateFormatterFormat|null $messageTimeType
      * @psalm-param SkipOnEmptyValue $skipOnEmpty
      * @psalm-param WhenType $when
      */
     public function __construct(
         ?string $format = null,
         private ?int $dateType = null,
+        private ?int $timeType = null,
         ?string $timeZone = null,
         ?string $locale = null,
         int|string|DateTimeInterface|null $min = null,
         int|string|DateTimeInterface|null $max = null,
         ?string $messageFormat = null,
         private ?int $messageDateType = null,
+        private ?int $messageTimeType = null,
         string $incorrectInputMessage = 'Invalid date value.',
         string $tooEarlyMessage = 'The value must be no early than {limit}.',
         string $tooLateMessage = 'The value must be no late than {limit}.',
@@ -68,13 +72,29 @@ final class Date extends BaseDate
     /**
      * @psalm-return IntlDateFormatterFormat|null
      */
+    public function getTimeType(): ?int
+    {
+        return $this->timeType;
+    }
+
+    /**
+     * @psalm-return IntlDateFormatterFormat|null
+     */
     public function getMessageDateType(): ?int
     {
         return $this->messageDateType;
     }
 
+    /**
+     * @psalm-return IntlDateFormatterFormat|null
+     */
+    public function getMessageTimeType(): ?int
+    {
+        return $this->messageTimeType;
+    }
+
     public function getHandler(): string
     {
-        return DateHandler::class;
+        return DateTimeHandler::class;
     }
 }