+++ src/Rule/Date/Time.php
@@ -16,25 +16,25 @@ use Yiisoft\Validator\WhenInterface;
  * @psalm-import-type WhenType from WhenInterface
  */
 #[Attribute(Attribute::TARGET_PROPERTY | Attribute::IS_REPEATABLE)]
-final class Date extends BaseDate
+final class Time extends BaseDate
 {
     /**
      * @psalm-param non-empty-string|null $timeZone
-     * @psalm-param IntlDateFormatterFormat|null $dateType
-     * @psalm-param IntlDateFormatterFormat|null $messageDateType
+     * @psalm-param IntlDateFormatterFormat|null $timeType
+     * @psalm-param IntlDateFormatterFormat|null $messageTimeType
      * @psalm-param SkipOnEmptyValue $skipOnEmpty
      * @psalm-param WhenType $when
      */
     public function __construct(
         ?string $format = null,
-        private ?int $dateType = null,
+        private ?int $timeType = null,
         ?string $timeZone = null,
         ?string $locale = null,
         int|string|DateTimeInterface|null $min = null,
         int|string|DateTimeInterface|null $max = null,
         ?string $messageFormat = null,
-        private ?int $messageDateType = null,
-        string $incorrectInputMessage = 'Invalid date value.',
+        private ?int $messageTimeType = null,
+        string $incorrectInputMessage = 'Invalid time value.',
         string $tooEarlyMessage = 'The value must be no early than {limit}.',
         string $tooLateMessage = 'The value must be no late than {limit}.',
         mixed $skipOnEmpty = null,
@@ -60,21 +60,21 @@ final class Date extends BaseDate
     /**
      * @psalm-return IntlDateFormatterFormat|null
      */
-    public function getDateType(): ?int
+    public function getTimeType(): ?int
     {
-        return $this->dateType;
+        return $this->timeType;
     }
 
     /**
      * @psalm-return IntlDateFormatterFormat|null
      */
-    public function getMessageDateType(): ?int
+    public function getMessageTimeType(): ?int
     {
-        return $this->messageDateType;
+        return $this->messageTimeType;
     }
 
     public function getHandler(): string
     {
-        return DateHandler::class;
+        return TimeHandler::class;
     }
 }